[chore] Refactoring tabscontentcomponent code
[chore] Fix the error Tabscontent on component

[chore] Fix remove unnecessary semi colon
